Fix a job that doesn't know to stop

+++ app/jobs/highlight_cleanup_job.rb
@@ -2,6 +2,8 @@ class HighlightCleanupJob < ApplicationJob
   queue_as :highlight_cleanup
 
   def perform(highlight_suggestion)
+    return if highlight_suggestion.nil?
+
     highlight_suggestion.destroy
   end
 end
